FAC-149:feature:Move jwt configs to common module
