+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/config/jwt/AuthTokenFilter.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.kit.config;
+package org.flickit.assessment.common.config.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
