+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/config/jwt/JwtTranslator.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.kit.config;
+package org.flickit.assessment.common.config.jwt;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
