+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/config/jwt/UserDetail.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.kit.config;
+package org.flickit.assessment.common.config.jwt;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
